#13 Refactor the custom hook to its own file
